add username to overview
