+++ app/src/main/res/layout/fragment_profile.xml
@@ -52,13 +52,13 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="5dp"
-                    android:text="username"
                     android:textAppearance="@style/TextAppearance.AppCompat.Body1"
                     android:textColor="#000000"
-                    android:textStyle="bold|italic"
+                    android:textStyle="bold"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
+                    app:layout_constraintTop_toTopOf="parent"
+                    tools:text="username" />
 
                 <TextView
                     android:id="@+id/tvCurrentBalance"
@@ -116,6 +116,7 @@
                     android:layout_width="367dp"
                     android:layout_height="21dp"
                     android:layout_marginTop="8dp"
+                    android:textAlignment="center"
                     android:textAppearance="@style/TextAppearance.AppCompat.Body1"
                     android:textSize="14sp"
                     app:layout_constraintEnd_toEndOf="parent"
@@ -128,6 +129,7 @@
                     android:layout_width="367dp"
                     android:layout_height="21dp"
                     android:layout_marginTop="4dp"
+                    android:textAlignment="center"
                     android:textAppearance="@style/TextAppearance.AppCompat.Body1"
                     android:textSize="14sp"
                     app:layout_constraintEnd_toEndOf="parent"
@@ -137,6 +139,23 @@
             </androidx.constraintlayout.widget.ConstraintLayout>
         </androidx.cardview.widget.CardView>
 
+        <androidx.cardview.widget.CardView
+            android:id="@+id/cvTopTokens"
+            android:layout_width="match_parent"
+            android:layout_height="300dp"
+            android:layout_margin="16dp"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="16dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/cvOverview">
+
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/rvTopTokens"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent">
+
+            </androidx.recyclerview.widget.RecyclerView>
+        </androidx.cardview.widget.CardView>
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rvTransactions"
@@ -159,7 +178,7 @@
             android:layout_height="wrap_content"
             android:layout_alignParentStart="true"
             android:layout_alignParentTop="true"
-            android:layout_marginTop="324dp"
+            android:layout_marginTop="16dp"
             android:text="Transaction History"
             android:textAppearance="@style/TextAppearance.AppCompat.Medium"
             android:textColor="#000000"
@@ -167,7 +186,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.038"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/cvOverview" />
+            app:layout_constraintTop_toBottomOf="@+id/cvTopTokens" />
 
         <ImageButton
             android:id="@+id/btnSettings"
